refactor(#725): rewire the router, tests, and patch targets to lookup.rowless
Rewire commit for the #725 move: every consumer of a moved name now imports it from lookup.rowless, with no re-export shim left behind. lookup/router.py (the program's only production import change outside the spine) now takes NONLIBRARY_RELEASE_SURFACED_STAT_KEY from lookup.rowless. tests/unit/test_nonlibrary_release_resolution.py retargets wholesale — its subject functions (_resolve_nonlibrary_release, _recover_track_credit) moved, so the module import, the five local _recover_track_credit imports, the two setattr(module, "_MAX_CREDIT_RECOVERY_FETCHES") sites, and the string patches on get_cached_release_id (x3) / set_cached_release_id (x3) / resolve_release_for_track (x2) all point at lookup.rowless, whose globals are what the moved kernel now reads. test_rowless_flag_observability.py takes NONLIBRARY_RELEASE_SURFACED_STAT_KEY and _make_rowless_item from lookup.rowless (x3 sites) while _library_miss_discogs_search stays an orchestrator import; test_rowless_enrichment.py, test_cached_track_safety_net.py, and test_orchestrator_helpers.py split ROWLESS_LIBRARY_ID out to lookup.rowless while their subjects (enrich_artwork_results, find_library_albums_with_cached_track, ROWLESS_NO_ALBUM_CONFIDENCE) stay on the orchestrator. Deliberately NOT retargeted: the _resolve_nonlibrary_release patches in test_album_title_fallback.py and test_lookup_nonlibrary_release.py stay on lookup.orchestrator because the code they intercept — search_compilations_for_track and the strategy callers under perform_lookup — still lives in the orchestrator and reads its module-global consumer import. Doc sweep found no file-path claim naming lookup/orchestrator.py as the home of a moved symbol (env-vars.md attaches the path only to the staying strategy functions; observability-rowless-flag.md names no file for the _make_rowless_item chokepoint), so no doc edits ride along.
